+++ Smithtry1000/VerticalLinesElement.cpp
@@ -1,47 +1,47 @@
 ﻿#include "VerticalLinesElement.h"
 
-void VerticalLinesElement::SetElectricalLength(double electricalLength)
+void VerticalLinesElement::SetElectricalLength(long double electricalLength)
 {
 	this->_electricalLength = electricalLength;
 }
 
-void VerticalLinesElement::SetMechanicalLength(double mechanicalLength)
+void VerticalLinesElement::SetMechanicalLength(long double mechanicalLength)
 {
 	this->_mechanicalLength = mechanicalLength;
 }
 
-void VerticalLinesElement::SetTheta(double theta)
+void VerticalLinesElement::SetTheta(long double theta)
 {
 	this->_theta = theta;
 }
 
-void VerticalLinesElement::SetLambda(double lambda)
+void VerticalLinesElement::SetLambda(long double lambda)
 {
 	this->_lambda = lambda;
 }
 
-double VerticalLinesElement::GetElectricalLength()
+long double VerticalLinesElement::GetElectricalLength()
 {
 	return this->_electricalLength;
 }
 
-double VerticalLinesElement::GetMechanicalLength()
+long double VerticalLinesElement::GetMechanicalLength()
 {
 	return this->_mechanicalLength;
 }
 
-double VerticalLinesElement::GetTheta()
+long double VerticalLinesElement::GetTheta()
 {
 	return this->_theta;
 }
 
-double VerticalLinesElement::GetLambda()
+long double VerticalLinesElement::GetLambda()
 {
 	return this->_lambda;
 }
 
-VerticalLinesElement::VerticalLinesElement(mode mode, double value, double frequency, Point point, map<chartMode, tuple<double, double>> chartParameters,
-	map<parameterMode, Complex> parameters, double electricalLength, double mechanicalLength, double theta, double lambda) :
+VerticalLinesElement::VerticalLinesElement(mode mode, long double value, long double frequency, Point point, map<chartMode, tuple<long double, long double>> chartParameters,
+	map<parameterMode, Complex> parameters, long double electricalLength, long double mechanicalLength, long double theta, long double lambda) :
 	Element(mode, value, frequency, point, chartParameters, parameters)
 {
 	this->SetElectricalLength(electricalLength);
